+++ tortoise/log.py
@@ -1,0 +1,4 @@
+import logging
+
+logger = logging.getLogger("tortoise")
+db_client_logger = logging.getLogger("tortoise.db_client")